+++ src/components/FeedList/FeedList.store.js
@@ -8,7 +8,7 @@ export const MODULE_NAME = 'FeedList'
 export const STORE_FETCH_POSTS_PARAM = `${MODULE_NAME}/STORE_FETCH_POSTS_PARAM`
 
 // actions
-export function fetchPosts ({ context, slug, sortBy, offset, search, filter, topic }) {
+export function fetchPosts ({ afterTime, beforeTime, context, filter, offset, order, search, slug, sortBy, topic }) {
   var query, extractModel, getItems
 
   if (context === 'groups') {
@@ -28,10 +28,13 @@ export function fetchPosts ({ context, slug, sortBy, offset, search, filter, top
     graphql: {
       query,
       variables: {
+        afterTime,
+        beforeTime,
+        context,
         filter,
         first: 20,
         offset,
-        context,
+        order,
         search,
         slug,
         sortBy,
@@ -49,10 +52,13 @@ export function fetchPosts ({ context, slug, sortBy, offset, search, filter, top
 }
 
 const groupQuery = `query (
+  $afterTime: Date,
+  $beforeTime: Date,
   $boundingBox: [PointInput],
   $filter: String,
   $first: Int,
   $offset: Int,
+  $order: String,
   $search: String,
   $slug: String,
   $sortBy: String,
@@ -76,12 +82,15 @@ const groupQuery = `query (
 }`
 
 const postsQuery = `query (
+  $afterTime: Date,
+  $beforeTime: Date,
   $boundingBox: [PointInput],
+  $context: String,
   $filter: String,
   $first: Int,
   $groupSlugs: [String],
   $offset: Int,
-  $context: String,
+  $order: String,
   $search: String,
   $sortBy: String,
   $topic: ID,