+++ src/routes/Events/index.js
@@ -1,0 +1,4 @@
+import component from './Events'
+import connector from './Events.connector'
+
+export default connector(component)