+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Events view now show Upcoming events in start time order and has an option to show Past events in descending start time order.
+
 ### Fixed
 - (Un)subscribing to a topic from a topic feed banner